DBG: fixed warning on x64

+++ src/dbg/_exports.cpp
@@ -887,7 +887,7 @@ extern "C" DLL_EXPORT duint _dbg_sendmessage(DBGMSG type, void* param1, void* pa
 
         duint animateInterval;
         if(BridgeSettingGetUint("Engine", "AnimateInterval", &animateInterval))
-            _dbg_setanimateinterval(animateInterval);
+            _dbg_setanimateinterval((unsigned int)animateInterval);
         else
             _dbg_setanimateinterval(50); // 20 commands per second
     }
